Rainbow
Added rainbow logic that appears it'll generate a nice transition of
color, once the color-delay math is worked out.

Given the update code, a single rainbow cycle takes average 85+(30*LEDs)
microseconds to complete.  In order to see a decently slow transition,
some time needs to be spent at each color step so that the changes can
be observed.